+++ packages/api/cms-api/src/common/enum/core-permission.enum.ts
@@ -10,4 +10,6 @@ export enum CorePermission {
     fileUploads = "fileUploads",
     dependencies = "dependencies",
     warnings = "warnings",
+    sitePreview = "sitePreview",
+    blockPreview = "blockPreview",
 }