+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/cli
 
+## 8.6.0
+
+### Minor Changes
+
+- 4c452a1: Add download-mitmproxy command
+
+## 8.5.2
+
+## 8.5.1
+
+## 8.5.0
+
 ## 8.4.2
 
 ## 8.4.1